Remove customizer opt in; Move opt in settings
