+++ theme/functions.php
@@ -259,7 +259,6 @@ require get_template_directory() . '/inc/customizer/colors.php';
 require get_template_directory() . '/inc/customizer/header-footer.php';
 require get_template_directory() . '/inc/customizer/layout.php';
 require get_template_directory() . '/inc/customizer/menu.php';
-require get_template_directory() . '/inc/customizer/fse-opt-in.php';
 
 /**
  * Custom menu walker
@@ -308,6 +307,5 @@ MaterialDesign\Theme\Customizer\Colors\setup();
 MaterialDesign\Theme\Customizer\Header_Footer\setup();
 MaterialDesign\Theme\Customizer\Layout\setup();
 MaterialDesign\Theme\Customizer\Menu\setup();
-MaterialDesign\Theme\Customizer\OptIn\setup();
 MaterialDesign\Theme\Widgets\setup();
 MaterialDesign\Theme\BlockEditor\setup();
